feat: move release out of sampo
