+++ crates/sampo-core/src/release.rs
@@ -1,6 +1,10 @@
 use crate::types::{Bump, CrateInfo, DependencyUpdate, Workspace};
-use crate::{changeset::ChangesetInfo, config::Config};
+use crate::{changeset::ChangesetInfo, config::Config, detect_changesets_dir, load_changesets, discover_workspace, detect_github_repo_slug_with_config, enrich_changeset_message, get_commit_hash_for_path};
+use rustc_hash::FxHashSet;
 use std::collections::{BTreeMap, BTreeSet};
+use std::fs;
+use std::io;
+use std::path::{Component, Path};
 
 /// Format dependency updates for changelog display
 ///
@@ -256,6 +260,680 @@ pub fn detect_fixed_dependency_policy_packages(
     }
 
     result
+}
+
+/// Type alias for initial bumps computation result
+type InitialBumpsResult = (
+    BTreeMap<String, Bump>,                // bump_by_pkg
+    BTreeMap<String, Vec<(String, Bump)>>, // messages_by_pkg
+    BTreeSet<std::path::PathBuf>,          // used_paths
+);
+
+/// Type alias for release plan
+type ReleasePlan = Vec<(String, String, String)>; // (name, old_version, new_version)
+
+/// Main release function that can be called from CLI or other interfaces
+pub fn run_release(root: &std::path::Path, dry_run: bool) -> io::Result<()> {
+    let ws = discover_workspace(root).map_err(io::Error::other)?;
+    let cfg = Config::load(&ws.root).map_err(io::Error::other)?;
+
+    // Validate fixed dependencies configuration
+    validate_fixed_dependencies(&cfg, &ws).map_err(io::Error::other)?;
+
+    let changesets_dir = detect_changesets_dir(&ws.root);
+    let changesets = load_changesets(&changesets_dir)?;
+    if changesets.is_empty() {
+        println!(
+            "No changesets found in {}",
+            ws.root.join(".sampo").join("changesets").display()
+        );
+        return Ok(());
+    }
+
+    // Compute initial bumps from changesets
+    let (mut bump_by_pkg, mut messages_by_pkg, used_paths) =
+        compute_initial_bumps(&changesets, &ws, &cfg)?;
+
+    if bump_by_pkg.is_empty() {
+        println!("No applicable packages found in changesets.");
+        return Ok(());
+    }
+
+    // Build dependency graph and apply cascading logic
+    let dependents = build_dependency_graph(&ws);
+    apply_dependency_cascade(&mut bump_by_pkg, &dependents, &cfg);
+    apply_linked_dependencies(&mut bump_by_pkg, &cfg);
+
+    // Prepare and validate release plan
+    let releases = prepare_release_plan(&bump_by_pkg, &ws)?;
+    if releases.is_empty() {
+        println!("No matching workspace crates to release.");
+        return Ok(());
+    }
+
+    print_release_plan(&releases);
+
+    if dry_run {
+        println!("Dry-run: no files modified, no tags created.");
+        return Ok(());
+    }
+
+    // Apply changes
+    apply_releases(&releases, &ws, &mut messages_by_pkg, &changesets, &cfg)?;
+
+    // Clean up
+    cleanup_consumed_changesets(used_paths)?;
+
+    Ok(())
+}
+
+/// Compute initial bumps from changesets and collect messages
+fn compute_initial_bumps(
+    changesets: &[ChangesetInfo],
+    ws: &Workspace,
+    cfg: &Config,
+) -> io::Result<InitialBumpsResult> {
+    let mut bump_by_pkg: BTreeMap<String, Bump> = BTreeMap::new();
+    let mut messages_by_pkg: BTreeMap<String, Vec<(String, Bump)>> = BTreeMap::new();
+    let mut used_paths: BTreeSet<std::path::PathBuf> = BTreeSet::new();
+
+    // Resolve GitHub repo slug once if available (config, env or origin remote)
+    let repo_slug = detect_github_repo_slug_with_config(&ws.root, cfg.github_repository.as_deref());
+    let github_token = std::env::var("GITHUB_TOKEN")
+        .ok()
+        .or_else(|| std::env::var("GH_TOKEN").ok());
+
+    for cs in changesets {
+        for pkg in &cs.packages {
+            used_paths.insert(cs.path.clone());
+            bump_by_pkg
+                .entry(pkg.clone())
+                .and_modify(|b| {
+                    if cs.bump > *b {
+                        *b = cs.bump;
+                    }
+                })
+                .or_insert(cs.bump);
+
+            // Enrich message with commit info and acknowledgments
+            let commit_hash = get_commit_hash_for_path(&ws.root, &cs.path);
+            let enriched = if let Some(hash) = commit_hash {
+                enrich_changeset_message(
+                    &cs.message,
+                    &hash,
+                    &ws.root,
+                    repo_slug.as_deref(),
+                    github_token.as_deref(),
+                    cfg.changelog_show_commit_hash,
+                    cfg.changelog_show_acknowledgments,
+                )
+            } else {
+                cs.message.clone()
+            };
+
+            messages_by_pkg
+                .entry(pkg.clone())
+                .or_default()
+                .push((enriched, cs.bump));
+        }
+    }
+
+    Ok((bump_by_pkg, messages_by_pkg, used_paths))
+}
+
+/// Build reverse dependency graph: dep -> set of dependents
+fn build_dependency_graph(ws: &Workspace) -> BTreeMap<String, BTreeSet<String>> {
+    let mut dependents: BTreeMap<String, BTreeSet<String>> = BTreeMap::new();
+    for c in &ws.members {
+        for dep in &c.internal_deps {
+            dependents
+                .entry(dep.clone())
+                .or_default()
+                .insert(c.name.clone());
+        }
+    }
+    dependents
+}
+
+/// Apply dependency cascade logic and fixed dependency groups
+fn apply_dependency_cascade(
+    bump_by_pkg: &mut BTreeMap<String, Bump>,
+    dependents: &BTreeMap<String, BTreeSet<String>>,
+    cfg: &Config,
+) {
+    // Helper function to find which fixed group a package belongs to, if any
+    let find_fixed_group = |pkg_name: &str| -> Option<usize> {
+        cfg.fixed_dependencies
+            .iter()
+            .position(|group| group.contains(&pkg_name.to_string()))
+    };
+
+    let mut queue: Vec<String> = bump_by_pkg.keys().cloned().collect();
+    let mut seen: BTreeSet<String> = queue.iter().cloned().collect();
+
+    while let Some(changed) = queue.pop() {
+        let changed_bump = bump_by_pkg.get(&changed).copied().unwrap_or(Bump::Patch);
+
+        // 1. Handle normal dependency relationships (unchanged → dependent)
+        if let Some(deps) = dependents.get(&changed) {
+            for dep_name in deps {
+                // Determine bump level for this dependent
+                let dependent_bump = if find_fixed_group(dep_name).is_some() {
+                    // Fixed dependencies: same bump level as the dependency
+                    changed_bump
+                } else {
+                    // Normal dependencies: at least patch
+                    Bump::Patch
+                };
+
+                let entry = bump_by_pkg
+                    .entry(dep_name.clone())
+                    .or_insert(dependent_bump);
+                // If already present, keep the higher bump
+                if *entry < dependent_bump {
+                    *entry = dependent_bump;
+                }
+                if !seen.contains(dep_name) {
+                    queue.push(dep_name.clone());
+                    seen.insert(dep_name.clone());
+                }
+            }
+        }
+
+        // 2. Handle fixed dependency groups (bidirectional)
+        if let Some(group_idx) = find_fixed_group(&changed) {
+            // All packages in the same fixed group should bump together
+            for group_member in &cfg.fixed_dependencies[group_idx] {
+                if group_member != &changed {
+                    let entry = bump_by_pkg
+                        .entry(group_member.clone())
+                        .or_insert(changed_bump);
+                    // If already present, keep the higher bump
+                    if *entry < changed_bump {
+                        *entry = changed_bump;
+                    }
+                    if !seen.contains(group_member) {
+                        queue.push(group_member.clone());
+                        seen.insert(group_member.clone());
+                    }
+                }
+            }
+        }
+    }
+}
+
+/// Apply linked dependencies logic: highest bump level to affected packages only
+fn apply_linked_dependencies(bump_by_pkg: &mut BTreeMap<String, Bump>, cfg: &Config) {
+    for group in &cfg.linked_dependencies {
+        // Check if any package in this group has been bumped
+        let mut group_has_bumps = false;
+        let mut highest_bump = Bump::Patch;
+
+        // First pass: find the highest bump level in the group among affected packages
+        for group_member in group {
+            if let Some(&member_bump) = bump_by_pkg.get(group_member) {
+                group_has_bumps = true;
+                if member_bump > highest_bump {
+                    highest_bump = member_bump;
+                }
+            }
+        }
+
+        // If any package in the group is being bumped, apply highest bump to affected packages only
+        if group_has_bumps {
+            // Apply the highest bump level to packages that are already being bumped
+            // (either directly affected or through dependency cascade)
+            for group_member in group {
+                if bump_by_pkg.contains_key(group_member) {
+                    // Only update if the current bump is lower than the group's highest bump
+                    let current_bump = bump_by_pkg
+                        .get(group_member)
+                        .copied()
+                        .unwrap_or(Bump::Patch);
+                    if highest_bump > current_bump {
+                        bump_by_pkg.insert(group_member.clone(), highest_bump);
+                    }
+                }
+            }
+        }
+    }
+}
+
+/// Prepare the release plan by matching bumps to workspace members
+fn prepare_release_plan(
+    bump_by_pkg: &BTreeMap<String, Bump>,
+    ws: &Workspace,
+) -> io::Result<ReleasePlan> {
+    // Map crate name -> CrateInfo for quick lookup
+    let mut by_name: BTreeMap<String, &CrateInfo> = BTreeMap::new();
+    for c in &ws.members {
+        by_name.insert(c.name.clone(), c);
+    }
+
+    let mut releases: Vec<(String, String, String)> = Vec::new(); // (name, old_version, new_version)
+    for (name, bump) in bump_by_pkg {
+        if let Some(info) = by_name.get(name) {
+            let old = if info.version.is_empty() {
+                "0.0.0".to_string()
+            } else {
+                info.version.clone()
+            };
+
+            let newv = bump_version(&old, *bump).unwrap_or_else(|_| old.clone());
+
+            releases.push((name.clone(), old, newv));
+        }
+    }
+
+    Ok(releases)
+}
+
+/// Print the planned releases
+fn print_release_plan(releases: &ReleasePlan) {
+    println!("Planned releases:");
+    for (name, old, newv) in releases {
+        println!("  {name}: {old} -> {newv}");
+    }
+}
+
+/// Apply all releases: update manifests and changelogs
+fn apply_releases(
+    releases: &ReleasePlan,
+    ws: &Workspace,
+    messages_by_pkg: &mut BTreeMap<String, Vec<(String, Bump)>>,
+    changesets: &[ChangesetInfo],
+    cfg: &Config,
+) -> io::Result<()> {
+    // Build lookup maps
+    let mut by_name: BTreeMap<String, &CrateInfo> = BTreeMap::new();
+    for c in &ws.members {
+        by_name.insert(c.name.clone(), c);
+    }
+
+    let mut new_version_by_name: BTreeMap<String, String> = BTreeMap::new();
+    for (name, _old, newv) in releases {
+        new_version_by_name.insert(name.clone(), newv.clone());
+    }
+
+    // Build releases map for dependency explanations
+    let releases_map: BTreeMap<String, (String, String)> = releases
+        .iter()
+        .map(|(name, old, new)| (name.clone(), (old.clone(), new.clone())))
+        .collect();
+
+    // Use unified function to detect all dependency explanations
+    let dependency_explanations =
+        detect_all_dependency_explanations(changesets, ws, cfg, &releases_map);
+
+    // Merge dependency explanations into existing messages
+    for (pkg_name, explanations) in dependency_explanations {
+        messages_by_pkg
+            .entry(pkg_name)
+            .or_default()
+            .extend(explanations);
+    }
+
+    // Apply updates for each release
+    for (name, old, newv) in releases {
+        let info = by_name.get(name.as_str()).unwrap();
+        let manifest_path = info.path.join("Cargo.toml");
+        let text = fs::read_to_string(&manifest_path)?;
+
+        // Update manifest versions
+        let (updated, _dep_updates) =
+            update_manifest_versions(&text, Some(newv.as_str()), ws, &new_version_by_name)?;
+        fs::write(&manifest_path, updated)?;
+
+        let messages = messages_by_pkg.get(name).cloned().unwrap_or_default();
+        update_changelog(&info.path, name, old, newv, &messages)?;
+    }
+
+    Ok(())
+}
+
+/// Clean up consumed changeset files
+fn cleanup_consumed_changesets(used_paths: BTreeSet<std::path::PathBuf>) -> io::Result<()> {
+    for p in used_paths {
+        let _ = fs::remove_file(p);
+    }
+    println!("Removed consumed changesets.");
+    Ok(())
+}
+
+/// Bump a semver version string
+pub fn bump_version(old: &str, bump: Bump) -> Result<String, String> {
+    let mut parts = old
+        .split('.')
+        .map(|s| s.parse::<u64>().unwrap_or(0))
+        .collect::<Vec<_>>();
+    while parts.len() < 3 {
+        parts.push(0);
+    }
+    let (maj, min, pat) = (parts[0], parts[1], parts[2]);
+    let (maj, min, pat) = match bump {
+        Bump::Patch => (maj, min, pat + 1),
+        Bump::Minor => (maj, min + 1, 0),
+        Bump::Major => (maj + 1, 0, 0),
+    };
+    Ok(format!("{maj}.{min}.{pat}"))
+}
+
+/// Update a crate manifest, setting the crate version (if provided) and retargeting
+/// internal dependency version requirements to the latest planned versions.
+/// Returns the updated TOML string along with a list of (dep_name, new_version) applied.
+pub fn update_manifest_versions(
+    input: &str,
+    new_pkg_version: Option<&str>,
+    ws: &Workspace,
+    new_version_by_name: &BTreeMap<String, String>,
+) -> io::Result<(String, Vec<(String, String)>)> {
+    let mut value: toml::Value = input
+        .parse()
+        .map_err(|e| io::Error::other(format!("invalid Cargo.toml: {e}")))?;
+
+    if let Some(v) = new_pkg_version
+        && let Some(pkg) = value.get_mut("package").and_then(toml::Value::as_table_mut)
+    {
+        pkg.insert("version".into(), toml::Value::String(v.to_string()));
+    }
+
+    let workspace_crates: BTreeSet<String> = ws.members.iter().map(|c| c.name.clone()).collect();
+    let mut applied: Vec<(String, String)> = Vec::new();
+
+    // helper to try update one dependency entry
+    fn update_dep_entry(
+        key: &str,
+        entry: &mut toml::Value,
+        workspace_crates: &BTreeSet<String>,
+        new_version_by_name: &BTreeMap<String, String>,
+        crate_dirs: &BTreeMap<String, std::path::PathBuf>,
+        base_dir: &std::path::Path,
+    ) -> Option<(String, String)> {
+        match entry {
+            toml::Value::String(ver) => {
+                // If the key itself matches a workspace crate with a new version, update string
+                if let Some(newv) = new_version_by_name.get(key)
+                    && workspace_crates.contains(key)
+                {
+                    *ver = newv.clone();
+                    return Some((key.to_string(), newv.clone()));
+                }
+            }
+            toml::Value::Table(tbl) => {
+                // Determine the real crate name: key or overridden via 'package'
+                let mut real_name = key.to_string();
+                if let Some(toml::Value::String(pkg_name)) = tbl.get("package") {
+                    real_name = pkg_name.clone();
+                }
+
+                // If path points to a workspace crate, prefer that crate's name
+                if let Some(toml::Value::String(path_str)) = tbl.get("path") {
+                    let dep_path = clean_path_like(&base_dir.join(path_str));
+                    if let Some(name) = crate_name_by_path(crate_dirs, &dep_path) {
+                        real_name = name;
+                    }
+                }
+
+                // Skip pure workspace deps (managed at workspace level)
+                if matches!(tbl.get("workspace"), Some(toml::Value::Boolean(true))) {
+                    return None;
+                }
+
+                if let Some(newv) = new_version_by_name.get(&real_name)
+                    && workspace_crates.contains(&real_name)
+                {
+                    tbl.insert("version".into(), toml::Value::String(newv.clone()));
+                    return Some((real_name, newv.clone()));
+                }
+            }
+            _ => {}
+        }
+        None
+    }
+
+    // Build helper maps for path resolution
+    let mut crate_dirs: BTreeMap<String, std::path::PathBuf> = BTreeMap::new();
+    for c in &ws.members {
+        crate_dirs.insert(c.name.clone(), c.path.clone());
+    }
+
+    // Resolve manifest base_dir from package.name
+    let current_crate_name = value
+        .get("package")
+        .and_then(toml::Value::as_table)
+        .and_then(|t| t.get("name"))
+        .and_then(toml::Value::as_str)
+        .unwrap_or("")
+        .to_string();
+    let base_dir = ws
+        .members
+        .iter()
+        .find(|c| c.name == current_crate_name)
+        .map(|c| c.path.as_path().to_path_buf())
+        .unwrap_or_else(|| std::path::PathBuf::from("."));
+
+    // Update dependencies across dependency sections
+    for section in ["dependencies", "dev-dependencies", "build-dependencies"] {
+        if let Some(t) = value.get_mut(section).and_then(toml::Value::as_table_mut) {
+            // Clone keys to avoid borrow issues while mutating
+            let keys: Vec<String> = t.keys().cloned().collect();
+            for dep_key in keys {
+                if let Some(entry) = t.get_mut(&dep_key)
+                    && let Some((dep_name, ver)) = update_dep_entry(
+                        &dep_key,
+                        entry,
+                        &workspace_crates,
+                        new_version_by_name,
+                        &crate_dirs,
+                        &base_dir,
+                    )
+                {
+                    applied.push((dep_name, ver));
+                }
+            }
+        }
+    }
+
+    // Also handle table-style per-dependency sections like [dependencies.foo]
+    // toml::Value already represents those as entries in the tables above, so no extra work.
+
+    let out = toml::to_string(&value)
+        .map_err(|e| io::Error::other(format!("failed to serialize Cargo.toml: {e}")))?;
+    Ok((out, applied))
+}
+
+fn crate_name_by_path(
+    crate_dirs: &BTreeMap<String, std::path::PathBuf>,
+    dep_path: &Path,
+) -> Option<String> {
+    let cleaned = clean_path_like(dep_path);
+    for (name, p) in crate_dirs {
+        if clean_path_like(p) == cleaned {
+            return Some(name.clone());
+        }
+    }
+    None
+}
+
+fn clean_path_like(p: &std::path::Path) -> std::path::PathBuf {
+    let mut out = std::path::PathBuf::new();
+    for c in p.components() {
+        match c {
+            Component::CurDir => {}
+            Component::ParentDir => {
+                if !matches!(
+                    out.components().next_back(),
+                    Some(Component::RootDir | Component::Prefix(_))
+                ) {
+                    out.pop();
+                }
+            }
+            _ => out.push(c),
+        }
+    }
+    out
+}
+
+fn update_changelog(
+    crate_dir: &Path,
+    package: &str,
+    old_version: &str,
+    new_version: &str,
+    entries: &[(String, Bump)],
+) -> io::Result<()> {
+    let path = crate_dir.join("CHANGELOG.md");
+    let existing = if path.exists() {
+        fs::read_to_string(&path)?
+    } else {
+        String::new()
+    };
+    let mut body = existing.trim_start_matches('\u{feff}').to_string();
+    // Remove existing top package header if present
+    let package_header = format!("# {}", package);
+    if body.starts_with(&package_header) {
+        if let Some(idx) = body.find('\n') {
+            body = body[idx + 1..].to_string();
+        } else {
+            body.clear();
+        }
+    }
+
+    // Parse and merge the current top section only if it's an unpublished section.
+    // Heuristic: if the top section header equals the current (old) version, it is published
+    // and must be preserved. Otherwise, treat it as in-progress and merge its bullets.
+    let mut merged_major: Vec<String> = Vec::new();
+    let mut merged_minor: Vec<String> = Vec::new();
+    let mut merged_patch: Vec<String> = Vec::new();
+
+    // helper to push without duplicates (preserve append order)
+    let push_unique = |list: &mut Vec<String>, msg: &str| {
+        if !list.iter().any(|m| m == msg) {
+            list.push(msg.to_string());
+        }
+    };
+
+    // Collect new entries
+    for (msg, bump) in entries {
+        match bump {
+            Bump::Major => push_unique(&mut merged_major, msg),
+            Bump::Minor => push_unique(&mut merged_minor, msg),
+            Bump::Patch => push_unique(&mut merged_patch, msg),
+        }
+    }
+
+    // If body starts with a previous top section (## ...), inspect its header.
+    // If header == old_version => preserve it (do not merge or strip).
+    // Else => parse and merge its bullets, then strip that section.
+    let trimmed = body.trim_start();
+    if trimmed.starts_with("## ") {
+        // Extract first header line text
+        let mut lines_iter = trimmed.lines();
+        let header_line = lines_iter.next().unwrap_or("").trim();
+        let header_text = header_line.trim_start_matches("## ").trim();
+
+        let is_published_top = header_text == old_version;
+
+        if !is_published_top {
+            // Determine the extent of the first section in 'trimmed'
+            let after_header_offset = header_line.len();
+            let rest_after_header = &trimmed[after_header_offset..];
+            // Find next section marker starting at a new line
+            let next_rel = rest_after_header.find("\n## ");
+            let (section_text, remaining) = match next_rel {
+                Some(pos) => {
+                    let end = after_header_offset + pos + 1; // include leading newline
+                    (&trimmed[..end], &trimmed[end..])
+                }
+                None => (trimmed, ""),
+            };
+
+            let mut current = None::<&str>;
+            for line in section_text.lines() {
+                let t = line.trim();
+                if t.eq_ignore_ascii_case("### Major changes") {
+                    current = Some("major");
+                    continue;
+                } else if t.eq_ignore_ascii_case("### Minor changes") {
+                    current = Some("minor");
+                    continue;
+                } else if t.eq_ignore_ascii_case("### Patch changes") {
+                    current = Some("patch");
+                    continue;
+                }
+                if t.starts_with("- ") {
+                    let msg = t.trim_start_matches("- ").trim();
+                    match current {
+                        Some("major") => push_unique(&mut merged_major, msg),
+                        Some("minor") => push_unique(&mut merged_minor, msg),
+                        Some("patch") => push_unique(&mut merged_patch, msg),
+                        _ => {}
+                    }
+                }
+            }
+
+            body = remaining.to_string();
+        }
+    }
+
+    // Build new aggregated top section
+    let mut section = String::new();
+    section.push_str(&format!("# {}\n\n", package));
+    section.push_str(&format!("## {}\n\n", new_version));
+
+    if !merged_major.is_empty() {
+        section.push_str("### Major changes\n\n");
+        for msg in &merged_major {
+            section.push_str("- ");
+            section.push_str(msg);
+            section.push('\n');
+        }
+        section.push('\n');
+    }
+    if !merged_minor.is_empty() {
+        section.push_str("### Minor changes\n\n");
+        for msg in &merged_minor {
+            section.push_str("- ");
+            section.push_str(msg);
+            section.push('\n');
+        }
+        section.push('\n');
+    }
+    if !merged_patch.is_empty() {
+        section.push_str("### Patch changes\n\n");
+        for msg in &merged_patch {
+            section.push_str("- ");
+            section.push_str(msg);
+            section.push('\n');
+        }
+        section.push('\n');
+    }
+
+    let combined = if body.trim().is_empty() {
+        section
+    } else {
+        format!("{}{}", section, body)
+    };
+    fs::write(&path, combined)
+}
+
+/// Validate fixed dependencies configuration against the workspace
+fn validate_fixed_dependencies(cfg: &Config, ws: &Workspace) -> Result<(), String> {
+    let workspace_packages: FxHashSet<String> = ws.members.iter().map(|c| c.name.clone()).collect();
+
+    for (group_idx, group) in cfg.fixed_dependencies.iter().enumerate() {
+        for package in group {
+            if !workspace_packages.contains(package) {
+                let available_packages: Vec<String> = workspace_packages.iter().cloned().collect();
+                return Err(format!(
+                    "Package '{}' in fixed dependency group {} does not exist in the workspace. Available packages: [{}]",
+                    package,
+                    group_idx + 1,
+                    available_packages.join(", ")
+                ));
+            }
+        }
+    }
+    Ok(())
 }
 
 #[cfg(test)]
